Update forms.py, models.py, and views.py
Got rid of recurrence_frequency from the model and form. Modified complete_task so that it would call create_recurring_task if the task had recurring. create_recurring_task creates the next task. Calculate_next_due_date is used to calculate the next due date of the new task depending on the pattern.